pyproject.toml
[project]
name = "stt-finetuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.8.4",
    "evaluate>=0.4.6",
    "jiwer>=4.0.0",
    "pyyaml>=6.0.3",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "transformers>=5.6.2",
]
